Add files via upload
This is Project is a problem set from the Harvard course' CS50.

And os basically a guess the sum up. The system will randomly pick 2 numbers, to sum up, and the user must say the right result.
If user guess it right, the user receives a cheer, if user guess it wrong, user gets a refer whistle.